@@ -1,7 +1,8 @@
 ﻿using System;
 using System.Data;
 using System.Data.Common;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using IDbConnection = EFCore.Toolkit.Abstractions.IDbConnection;
 
 namespace EFCore.Toolkit.Extensions
