@@ -1,10 +1,22 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
 
 namespace EFCore.Toolkit.Abstractions.Extensions
 {
     public static class ReadOnlyRepositoryExtensions
-    {
+    {  /// <summary>
+       ///     Finds entities with the given <paramref name="predicate"/>.
+       /// </summary>
+       /// <param name="predicate">The search predicate.</param>
+       /// <returns>A collection of entities matching the search predicate.</returns>
+        public static IEnumerable<T> FindBy<T>(this IReadOnlyRepository<T> repository, Expression<Func<T, bool>> predicate)
+        {
+            IEnumerable<T> query = repository.Get().Where(predicate).AsEnumerable();
+            return query;
+        }
+
         /// <summary>
         ///     Indicates whether an entity with the given primary key value exists.
         /// </summary>
