@@ -1,12 +1,11 @@
 ﻿using System;
-using System.Linq.Expressions;
 using System.Threading.Tasks;
-#if !NET40
-
-#endif
 
 namespace EFCore.Toolkit.Abstractions
-{
+{    
+     /// <summary>
+     /// IContext is the database-independant abstraction of a data access context.
+     /// </summary>
     public interface IContext : IDisposable
     {
         /// <summary>
@@ -17,33 +16,28 @@ namespace EFCore.Toolkit.Abstractions
 
         /// <summary>
         ///     Drops the underlying database.
-        ///     USE WITH CARE!
         /// </summary>
         void DropDatabase();
 
         /// <summary>
-        ///     Updates the given entity.
+        ///     Sets the state of <paramref name="entity"/> to given modified.
         /// </summary>
-        TEntity Edit<TEntity>(TEntity entity) where TEntity : class;
+        void SetStateModified<TEntity>(TEntity entity) where TEntity : class;
 
         /// <summary>
-        ///     Updates the given attached <paramref name="originalEntity" /> with the new entity <paramref name="updateEntity" />.
+        ///     Sets the state of <paramref name="entity"/> to given unchanged.
         /// </summary>
-        TEntity Edit<TEntity>(TEntity originalEntity, TEntity updateEntity) where TEntity : class;
-
-        TEntity Delete<TEntity>(TEntity entity) where TEntity : class;
+        void SetStateUnchanged<TEntity>(TEntity entity) where TEntity : class;
 
         /// <summary>
-        ///     Reverts changes in <paramref name="entity" />.
+        ///     Updates the properties of <paramref name="originalEntity" /> with values from <paramref name="updateEntity" />.
         /// </summary>
-        void UndoChanges<TEntity>(TEntity entity) where TEntity : class;
+        TEntity SetValues<TEntity>(TEntity originalEntity, TEntity updateEntity) where TEntity : class;
 
         /// <summary>
         ///     Modifies the properties with <paramref name="propertyNames" /> of given entity <paramref name="entity" />.
         /// </summary>
         void ModifyProperties<TEntity>(TEntity entity, params string[] propertyNames) where TEntity : class;
-
-        void LoadReferenced<TEntity, TProperty>(TEntity entity, Expression<Func<TEntity, TProperty>> navigationProperty) where TEntity : class where TProperty : class;
 
         /// <summary>
         ///     Saves all changes made in this context to the underlying database.
@@ -52,16 +46,15 @@ namespace EFCore.Toolkit.Abstractions
         /// <exception cref="System.InvalidOperationException">Thrown if the context has been disposed.</exception>
         ChangeSet SaveChanges();
 
-#if !NET40
         /// <summary>
         ///     Saves all changes made in this context to the underlying database.
         /// </summary>
         /// <returns>The number of objects written to the underlying database.</returns>
         /// <exception cref="System.InvalidOperationException">Thrown if the context has been disposed.</exception>
         Task<ChangeSet> SaveChangesAsync();
-#endif
 
         ITransaction BeginTransaction();
+
         void UseTransaction(ITransaction transaction);
     }
 
