@@ -1,14 +1,21 @@
 ﻿using EFCore.Toolkit.Abstractions.Auditing;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace EFCore.Toolkit.Auditing
 {
     public abstract class AuditEntityTypeConfiguration<TAuditEntity, TAuditKey> :
-        EntityTypeConfiguration<TAuditEntity> where TAuditEntity : class,
+        IEntityTypeConfiguration<TAuditEntity> where TAuditEntity : class,
         IAuditEntity<TAuditKey>
     {
-        public override void Configure(EntityTypeBuilder<TAuditEntity> entity)
+        protected virtual void Configure(EntityTypeBuilder<TAuditEntity> entity)
         {
+        }
+
+        void IEntityTypeConfiguration<TAuditEntity>.Configure(EntityTypeBuilder<TAuditEntity> entity)
+        {
+            this.Configure(entity);
+
             entity.HasKey(e => e.AuditId);
             entity.Property(e => e.AuditId).IsRequired();
             entity.Property(e => e.AuditDate).IsRequired();
