@@ -128,27 +128,26 @@ namespace EFCore.Toolkit.Extensions
             return entity;
         }
 
-#if !NETSTANDARD1_3 && !NETFX
         /// <summary>
         /// Returns the number of table rows per database table.
         /// </summary>
         public static List<TableRowCounts> GetTableRowCounts(this DbContext c)
         {
-            var rawSqlQuery = c.Query<TableRowCounts>().FromSql(
-                @"CREATE TABLE #counts
-                    (
-                        TableName varchar(255),
-                        TableRowCount int
-                    )
+            throw new NotImplementedException("GetTableRowCounts is currently not implemented");
+            ////var rawSqlQuery = c.Query<TableRowCounts>().FromSql(
+            ////    @"CREATE TABLE #counts
+            ////        (
+            ////            TableName varchar(255),
+            ////            TableRowCount int
+            ////        )
 
-                    EXEC sp_MSForEachTable @command1='INSERT #counts (TableName, TableRowCount) SELECT ''?'', COUNT(*) FROM ?'
-                    SELECT TableName, TableRowCount FROM #counts ORDER BY TableName, TableRowCount DESC
-                    DROP TABLE #counts");
+            ////        EXEC sp_MSForEachTable @command1='INSERT #counts (TableName, TableRowCount) SELECT ''?'', COUNT(*) FROM ?'
+            ////        SELECT TableName, TableRowCount FROM #counts ORDER BY TableName, TableRowCount DESC
+            ////        DROP TABLE #counts");
 
-            var tableCountResults = rawSqlQuery.ToList();
-            return tableCountResults;
+            ////var tableCountResults = rawSqlQuery.ToList();
+            ////return tableCountResults;
         }
-#endif
 
         public static IQueryable Set(this DbContext context, Type entityType)
         {
