@@ -8,6 +8,10 @@ using EFCore.Toolkit.Abstractions.Extensions;
 
 namespace EFCore.Toolkit
 {
+    /// <summary>
+    /// The in-memory representation of <seealso cref="IGenericRepository{T}"/>.
+    /// </summary>
+    /// <typeparam name="T">The entity type.</typeparam>
     public class InMemoryRepository<T> : IGenericRepository<T> where T : IIdentifiable
     {
         private readonly List<T> items;
@@ -26,36 +30,38 @@ namespace EFCore.Toolkit
             this.items.Clear();
         }
 
+        /// <inheritdoc />
         public ChangeSet Save()
         {
             return ChangeSet.Empty;
         }
 
+        /// <inheritdoc />
         public Task<ChangeSet> SaveAsync()
         {
             return Task.FromResult(ChangeSet.Empty);
         }
 
+        /// <inheritdoc />
         public IQueryable<T> Get()
         {
             return this.items.AsQueryable();
         }
 
+        /// <inheritdoc />
         public IEnumerable<T> GetAll()
         {
             return this.items;
         }
 
+        /// <inheritdoc />
         public T FindById(params object[] ids)
         {
-            return this.items.FirstOrDefault(); // TODO Wrong implementation
+            var intIds = ids.Select(i => int.Parse($"{i}"));
+            return this.items.SingleOrDefault(i => intIds.Contains(i.Id)); // TODO Test this implementation
         }
 
-        public IEnumerable<T> FindBy(Expression<Func<T, bool>> predicate)
-        {
-            return this.items;
-        }
-
+        /// <inheritdoc />
         public T Add(T entity)
         {
             entity.Id = this.items.GetNextId();
@@ -64,6 +70,7 @@ namespace EFCore.Toolkit
             return entity;
         }
 
+        /// <inheritdoc />
         public IEnumerable<T> AddRange(IEnumerable<T> entities)
         {
             var collection = entities.ToList();
@@ -75,6 +82,7 @@ namespace EFCore.Toolkit
             return collection;
         }
 
+        /// <inheritdoc />
         public T AddOrUpdate(T entity)
         {
             this.Remove(entity);
@@ -83,6 +91,7 @@ namespace EFCore.Toolkit
             return entity;
         }
 
+        /// <inheritdoc />
         public T Update(T entity)
         {
             this.items.Remove(entity);
@@ -91,7 +100,15 @@ namespace EFCore.Toolkit
             return entity;
         }
 
-        public T Update(T entity, T updateEntity)
+        /// <inheritdoc />
+        public void UpdateRange(IEnumerable<T> entities)
+        {
+            this.RemoveRange(entities);
+            this.AddRange(entities);
+        }
+
+        /// <inheritdoc />
+        public T SetValues(T entity, T updateEntity)
         {
             this.items.Remove(entity);
             this.items.Add(updateEntity);
@@ -99,6 +116,7 @@ namespace EFCore.Toolkit
             return updateEntity;
         }
 
+        /// <inheritdoc />
         public T UpdateProperties<TValue>(T entity, params Expression<Func<T, TValue>>[] propertyExpressions)
         {
             this.Remove(entity);
@@ -107,6 +125,7 @@ namespace EFCore.Toolkit
             return entity;
         }
 
+        /// <inheritdoc />
         public T UpdateProperty<TValue>(T entity, Expression<Func<T, TValue>> propertyExpression, TValue value)
         {
             this.Remove(entity);
@@ -115,15 +134,20 @@ namespace EFCore.Toolkit
             return entity;
         }
 
+        /// <inheritdoc />
         public T Remove(T entity)
         {
             this.items.Remove(entity);
-
             return entity;
         }
 
-        public void LoadReferenced<TEntity, TProperty>(TEntity entity, Expression<Func<TEntity, TProperty>> navigationProperty) where TEntity : class where TProperty : class
+        /// <inheritdoc />
+        public IEnumerable<T> RemoveRange(IEnumerable<T> entities)
         {
+            foreach (var entity in entities)
+            {
+                yield return this.Remove(entity);
+            }
         }
 
         public IContext Context { get; }
