@@ -1,16 +1,10 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
+using System.Threading;
 
 namespace EFCore.Toolkit.Testing
 {
-    /// <summary>
-    /// TestAsyncEnumerable&lt;T&gt; implements <seealso cref="IAsyncEnumerable&lt;T&gt;"/>
-    /// which is used to mock collections behind async queryables -> ToListAsync
-    /// 
-    /// Source: https://stackoverflow.com/questions/40476233/how-to-mock-an-async-repository-with-entity-framework-core
-    /// </summary>
-    /// <typeparam name="T">Entity type.</typeparam>
     public class TestAsyncEnumerable<T> : EnumerableQuery<T>, IAsyncEnumerable<T>, IQueryable<T>
     {
         public TestAsyncEnumerable(IEnumerable<T> enumerable)
@@ -23,11 +17,16 @@ namespace EFCore.Toolkit.Testing
         {
         }
 
-        public IAsyncEnumerator<T> GetEnumerator()
+        IQueryProvider IQueryable.Provider => new TestAsyncQueryProvider<T>(this);
+
+        public IAsyncEnumerator<T> GetAsyncEnumerator(CancellationToken cancellationToken = new CancellationToken())
         {
             return new TestAsyncEnumerator<T>(this.AsEnumerable().GetEnumerator());
         }
 
-        IQueryProvider IQueryable.Provider => new TestAsyncQueryProvider<T>(this);
+        public IAsyncEnumerator<T> GetEnumerator()
+        {
+            return this.GetAsyncEnumerator();
+        }
     }
 }
