@@ -1,12 +1,13 @@
 ﻿using EFCore.Toolkit;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using ToolkitSample.Model;
 
 namespace ToolkitSample.DataAccess.Context
 {
-    public class ApplicationSettingEntityTypeConfiguration : EntityTypeConfiguration<ApplicationSetting>
+    public class ApplicationSettingEntityTypeConfiguration : IEntityTypeConfiguration<ApplicationSetting>
     {
-        public override void Configure(EntityTypeBuilder<ApplicationSetting> entity)
+        public void Configure(EntityTypeBuilder<ApplicationSetting> entity)
         {
             entity.HasKey(d => d.Id);
             entity.Property(d => d.Id).ValueGeneratedNever();
