@@ -1,12 +1,13 @@
 ﻿using EFCore.Toolkit;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using ToolkitSample.Model.Auditing;
 
 namespace ToolkitSample.DataAccess.Context
 {
-    public class TestEntityEntityTypeConfiguration : EntityTypeConfiguration<TestEntity>
+    public class TestEntityEntityTypeConfiguration : IEntityTypeConfiguration<TestEntity>
     {
-        public override void Configure(EntityTypeBuilder<TestEntity> entity)
+        public void Configure(EntityTypeBuilder<TestEntity> entity)
         {
             entity.HasKey(e => e.TestEntityId);
         }
