@@ -1,5 +1,4 @@
-﻿#if NETSTANDARD
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
 using ToolkitSample.DataAccess.Context;
@@ -23,4 +22,3 @@ namespace ToolkitSample.DataAccess
         }
     }
 }
-#endif
